% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.moments.lin.r
\name{compute.moments.lin}
\alias{compute.moments.lin}
\title{A function to compute moments with a linear transect}
\usage{
compute.moments.lin(ab, cv, ht, q)
}
\arguments{
\item{ab}{vector of abundances}

\item{cv}{vector of coverages}

\item{ht}{vector of heights}

\item{q}{vector of orders}
}
\description{
A function to compute moments with a linear transect
}
