% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.spectrum.r
\name{compute.spectrum}
\alias{compute.spectrum}
\title{A function to compute global multifractal spectrum}
\usage{
compute.spectrum(mom)
}
\arguments{
\item{mom}{moments object returned by compute.moments.lin()}
}
\description{
A function to compute global multifractal spectrum
}
