% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derv.r
\name{derv}
\alias{derv}
\title{A function to compute local derivative}
\usage{
derv(vec, h)
}
\arguments{
\item{vec}{vector}

\item{h}{interval size}
}
\description{
A function to compute local derivative
}
\keyword{derivative}
