\name{ecomf-package}
\alias{ecomf-package}
\alias{ecomf}
\docType{package}
\title{
\packageTitle{ecomf}
}
\description{
\packageDescription{ecomf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ecomf}
\packageIndices{ecomf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ecomf}

Maintainer: \packageMaintainer{ecomf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
