% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.spectra.r
\name{local.spectra}
\alias{local.spectra}
\title{A function to compute local multifractal spectra}
\usage{
local.spectra(inp, sc = 0, smooth = 1)
}
\arguments{
\item{inp}{input object returned by compute.moments.lin()}

\item{sc}{vector of sample scales}

\item{smooth}{smoothing parameter}
}
\description{
A function to compute local multifractal spectra
}
