% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.spectra.r
\name{trunc.spectra}
\alias{trunc.spectra}
\title{A function to truncate anomalous "horns" from multifrcatal spectra}
\usage{
\method{trunc}{spectra}(mf)
}
\arguments{
\item{mf}{spectra object returned by local.spectra()}
}
\description{
A function to truncate anomalous "horns" from multifrcatal spectra
}
